+++ libai/inference/basic.py
@@ -123,7 +123,7 @@ class BasePipeline(metaclass=ABCMeta):
                 set it to `random` for quickly debugging by random initialized model
         """
         if mode == "libai":
-            from libai.models.utils.model_utils.base_loader import ModelLoaderLiBai
+            from libai.models.utils.model_loader.base_loader import ModelLoaderLiBai
 
             model_loader = ModelLoaderLiBai(libai_cfg_model, libai_cfg_model.cfg, model_path)
             model_loader.base_model_prefix_1 = None
