+++ libai/models/utils/model_loader/swin_loader.py
