+++ libai/models/utils/model_loader/vit_loader.py
