+++ tests/model_loader/test_swin_loader.py
