+++ tests/model_loader/test_swinv2_loader.py
