+++ tests/model_loader/test_vit_loader.py
